src/test/java/ru/netology/stats/StatsServiceTest.java
package ru.netology.stats;

import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.Test;

public class StatsServiceTest {

    @Test
    void statisticsSumAllSales() {
        StatsService service = new StatsService();
        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
        long expected = 180;
        int actual = (int) service.sumAllSales(sales);
        Assertions.assertEquals(actual, expected);
    }

    @Test
    void statisticsAverageSumSales() {
        StatsService service = new StatsService();
        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
        double expected = 15;
        double actual = service.averageSumSales(sales);
        Assertions.assertEquals(actual, expected);
    }

    @Test
    void statisticsMaxMonth() {
        StatsService service = new StatsService();
        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
        long expected = 8;
        int actual = (int) service.maxSales(sales);
        Assertions.assertEquals(actual, expected);
    }

    @Test
    void statisticsMinMonth() {
        StatsService service = new StatsService();
        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
        long expected = 9;
        int actual = (int) service.minSales(sales);
        Assertions.assertEquals(actual, expected);
    }

    @Test
    void statisticsSalesBelowAverage() {
        StatsService service = new StatsService();
        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
        long expected = 5;
        int actual = service.salesBelowAverage(sales);
        Assertions.assertEquals(actual, expected);
    }

    @Test
    void statisticsSalesAboveAverage() {
        StatsService service = new StatsService();
        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
        long expected = 5;
        int actual = service.salesBelowAverage(sales);
        Assertions.assertEquals(actual, expected);
    }
}
